pixel/requirements.txt
django==3.2.7
django-crispy-forms==1.12.0
Pillow==8.3.2
numpy==1.21.2
sorl-thumbnail==12.7.0